backend/auth_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pyotp
python-multipart
httpx
alembic
python-dotenv
